Automatic loading of translations works.
Translations are then being inserted into views using the "<?= f('WHAT') ?>" directive.
These directives are only present in layout.phtml so far.
The translation files are all the same now, except for the locales/cs/translation.json file, which is partially translated and can be used for testing.
To see the difference, just append "?lang=cs" to the URL and load the webpage. To go back to English, simply append "?lang=en" to one request.