The AVAILABLE_LANGUAGES constant is replaced with a property
This property is automatically filled in in the constructor of the LanguageHandler class.
The constructor scans the default locale directory and loads all available languages.
For this to work, the proper directory structure must be kept.
A new file is created in all locale folders, named ".LOCALENAME" and contains nothing but a name of the language as it should display in the language selection box.
This was needed because there isn't any other easy way to know what to display without loading any dictionaries of all the available languages.